Added content

+++ CPP/GameDev/FirstGame/FirstGame/Engine/Engine.h
@@ -10,7 +10,7 @@ class Engine {
 public:
 	static int SCREEN_WIDTH;
 	static int SCREEN_HEIGHT;
-	static double GetDT();
+	static float GetDT();
 
 	Engine();
 	~Engine();
@@ -27,8 +27,8 @@ public:
 private:
 	static GLFWwindow* window;
 
-	static double dt;
-	double lastTime;
+	static float dt;
+	float lastTime;
 	bool joystickUse;
 
 	Joystick joysticks[GLFW_JOYSTICK_LAST];
